Large refactor for performance increase Several changes have been made to the component for how it handles the path generation and management.
Before the canvas was sized to the webpage, it is no longer. It is now
sized to just the size needed to hold the path shape.  then a CSS
transform is applied to translate the canvas to the correct location.

Second performance was the elimination of extra redrawing.  The crux of
the problem is that the concepts don't know about the paths and the
paths don't know about the concepts, but depend on the concepts being
mounted to the DOM.  So a "channel" for communicating updates is now
used to allow concepts to self-register their location and for paths to
declare their need for them without causing extra renders.

So now, on load there is one DOM paint for the elements, and one canvas
paint per canvas to show the paths.

Some concerns exist whether this approach will work in the longer run,
so an SVG approach should probably be explored.

Infinite gratitude to @iHiD and @SleeplessByte for their help

Co-authored-by: Jeremy Walker <jez.walker@gmail.com>
Co-authored-by:
Derk-Jan Karrenbeld <derk-jan+github@karrenbeld.info>