<ConceptMap /> SVG, Highlighting, CSS, oh my! (#205)
* controller updated to new interface
now uses UserTrack::GenerateConceptStatusMapping to create
status mapping

* ancestors now highlight on concept hover

* Tidying up active and inactive concept states

* Single canvas approach changed to multi-canvas

* ConceptState -> ConceptStatus

* remove debug console messages

* Large refactor for performance increase
Several changes have been made to the component for how it handles the
path generation and management.

Before the canvas was sized to the webpage, it is no longer. It is now
sized to just the size needed to hold the path shape.  then a CSS
transform is applied to translate the canvas to the correct location.

Second performance was the elimination of extra redrawing.  The crux of
the problem is that the concepts don't know about the paths and the
paths don't know about the concepts, but depend on the concepts being
mounted to the DOM.  So a "channel" for communicating updates is now
used to allow concepts to self-register their location and for paths to
declare their need for them without causing extra renders.

So now, on load there is one DOM paint for the elements, and one canvas
paint per canvas to show the paths.

Some concerns exist whether this approach will work in the longer run,
so an SVG approach should probably be explored.

Infinite gratitude to @iHiD and @SleeplessByte for their help

Co-authored-by: Jeremy Walker <jez.walker@gmail.com>
Co-authored-by:
Derk-Jan Karrenbeld <derk-jan+github@karrenbeld.info>

* fix nullish coalescing for undefined

* fix webpage size handler from causign extra render

* add visibility handler

* MASSIVE REFACTOR
Now uses SVG like it should have all along. 🤯

types cleaned up, files cleaned up, still more to do for that I'm sure

* fix test import, let to const

* svg styling

* add tailwind negative zindexes

* add module boundary function return types

* Add render when custom font loads.
As it is, the rendering concept map is faster than the loading the
custom font "Poppins". When the font loads, the font is swapped causing
the concept's DOM width to change but this doesn't trigger a react
render.  This adds the useFontLoaded hook so it will trigger a reload
when the font load event is detected.

* Optimize all of the things!
- Memoization added to several aspects so that it only recalculates what
changes.
- Request animation frame wrappers added to redrawing calls.

- some type cleanup
- unused function cleanup

* fix controller

* Changes based on review from @iHiD
type ConceptPath is now ConceptPathProperties, which is computede by
`computePathProperties`. Led to a refactor since we don't need to maintain the
absolute coordinates for the svg path any more, and can just use the
size of the computed viewbox to determine the start and end based
on the coordinates of the concept boxes.

Also several CSS improvements.