Patch: Fix tooltip dispatch helper (#394)
When tabbing through the concepts, if tab action caused focus to the
tooltip (if the the html sent over the wire contained a focusable
element (e.g. a hyperlink <a />)) it would become stuck open because the
helper was sending the wrong action to the state reducer.

This fixes that behavior.