Prevent race conditions
Some models used the `after_save` and `after_update` callbacks, but these could lead to race conditions as the code called
within those methods could read an outdated version of the model as there was no guarantee the changes would have been committed.

This was e.g. problematic when dealing with a solution's published head tests status, which would not be updated due to the published
solution's published_iteration_id not having yet been updated.